+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
-import '../src/styles/globals.css';
-import BaseLayout from '../src/layouts/BaseLayout';
+import '../styles/globals.css';
+import BaseLayout from '../layouts/BaseLayout';
 
 function MyApp({ Component, pageProps }: AppProps) {
 	return (
